+++ rsbuild.config.ts
@@ -1,7 +1,16 @@
-import { defineConfig } from '@rsbuild/core';
+import { defineConfig, rspack } from '@rsbuild/core';
 import { pluginNodePolyfill } from '@rsbuild/plugin-node-polyfill';
 
 import { version } from './package.json';
+
+const nodeOnlyModules = /^proxy-agent$/;
+const unusedInBrowserBuiltins = ['os', 'zlib', 'util'];
+const builtinAliases = Object.fromEntries(
+    unusedInBrowserBuiltins.flatMap((m) => [
+        [m, false],
+        [`node:${m}`, false],
+    ]),
+);
 
 // eslint-disable-next-line import/no-default-export
 export default defineConfig({
@@ -38,7 +47,8 @@ export default defineConfig({
                     type: 'umd', // or 'umd', 'commonjs', etc.
                     name: 'Apify',
                 },
-                globalObject: 'this',
+                globalObject: 'globalThis',
+                asyncChunks: false,
             };
             config.optimization = {
                 ...config.optimization,
@@ -47,11 +57,18 @@ export default defineConfig({
                 splitChunks: false,
                 minimize: false,
             };
+            config.plugins = [...(config.plugins ?? []), new rspack.IgnorePlugin({ resourceRegExp: nodeOnlyModules })];
+            config.resolve = {
+                ...config.resolve,
+                alias: {
+                    ...config.resolve?.alias,
+                    ...builtinAliases,
+                },
+            };
             config.devtool = 'source-map';
         },
     },
-    // mode: 'production',
-    mode: 'development',
+    mode: 'production',
     // @apify/utilities dynamically imports `crypto` on missing `SubtleCrypto` (but browsers have it).
     plugins: [pluginNodePolyfill({ overrides: { crypto: false } })],
 });